requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
pyTelegramBotAPI==4.11.0
python-dotenv==1.0.0
requests==2.28.2
urllib3==1.26.15
